+++ emulinker/pom.xml
@@ -20,6 +20,13 @@
 
   <dependencies>
 
+    <!-- https://mvnrepository.com/artifact/io.dropwizard.metrics/metrics-core -->
+    <dependency>
+      <groupId>io.dropwizard.metrics</groupId>
+      <artifactId>metrics-core</artifactId>
+      <version>4.2.3</version>
+    </dependency>
+
     <dependency>
       <groupId>com.google.flogger</groupId>
       <artifactId>flogger</artifactId>
