+++ emulinker/src/main/java/org/emulinker/kaillera/controller/v086/V086ClientHandlerFactory.java
@@ -1,0 +1,9 @@
+package org.emulinker.kaillera.controller.v086;
+
+import dagger.assisted.AssistedFactory;
+import java.net.InetSocketAddress;
+
+@AssistedFactory
+public interface V086ClientHandlerFactory {
+    public V086ClientHandler create(InetSocketAddress remoteSocketAddress, V086Controller v086Controller);
+}
