+++ emulinker/src/main/java/org/emulinker/kaillera/controller/v086/action/PlayerDesynchAction.java
@@ -4,7 +4,7 @@ import com.google.common.flogger.FluentLogger;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 import org.emulinker.kaillera.controller.messaging.MessageFormatException;
-import org.emulinker.kaillera.controller.v086.V086Controller.V086ClientHandler;
+import org.emulinker.kaillera.controller.v086.V086ClientHandler;
 import org.emulinker.kaillera.controller.v086.protocol.GameChat_Notification;
 import org.emulinker.kaillera.model.event.*;
 import org.emulinker.util.EmuLang;
