+++ emulinker/src/main/java/org/emulinker/kaillera/controller/v086/action/V086Action.java
@@ -1,6 +1,6 @@
 package org.emulinker.kaillera.controller.v086.action;
 
-import org.emulinker.kaillera.controller.v086.V086Controller.V086ClientHandler;
+import org.emulinker.kaillera.controller.v086.V086ClientHandler;
 import org.emulinker.kaillera.controller.v086.protocol.V086Message;
 
 public interface V086Action<T extends V086Message> {
