+++ emulinker/src/main/java/org/emulinker/kaillera/controller/v086/action/V086GameEventHandler.java
@@ -1,6 +1,6 @@
 package org.emulinker.kaillera.controller.v086.action;
 
-import org.emulinker.kaillera.controller.v086.V086Controller.V086ClientHandler;
+import org.emulinker.kaillera.controller.v086.V086ClientHandler;
 import org.emulinker.kaillera.model.event.GameEvent;
 
 public interface V086GameEventHandler<T extends GameEvent> {
