+++ emulinker/src/main/java/org/emulinker/net/UDPServer.java
@@ -1,5 +1,7 @@
 package org.emulinker.net;
 
+import com.codahale.metrics.MetricRegistry;
+import com.codahale.metrics.Timer;
 import com.google.common.flogger.FluentLogger;
 import java.io.IOException;
 import java.net.*;
@@ -36,12 +38,13 @@ public abstract class UDPServer implements Executable {
   private boolean isRunning = false;
   private boolean stopFlag = false;
 
-  public UDPServer() {
-    this(true);
-  }
+  protected final Timer serverToClientRequests;
 
-  public UDPServer(boolean shutdownOnExit) {
+  public UDPServer(boolean shutdownOnExit, MetricRegistry metrics) {
     if (shutdownOnExit) Runtime.getRuntime().addShutdownHook(new ShutdownThread());
+
+    this.serverToClientRequests =
+        metrics.timer(MetricRegistry.name(UDPServer.class, "clientToServerRequests"));
   }
 
   public int getBindPort() {
@@ -180,6 +183,7 @@ public abstract class UDPServer implements Executable {
           // https://stackoverflow.com/a/61267496/2875073
           ((Buffer) buffer).flip();
           //					logger.atFine().log("receive("+EmuUtil.dumpBuffer(buffer, false)+")");
+          // TODO(nue): time this
           handleReceived(buffer, fromSocketAddress);
           releaseBuffer(buffer);
         } catch (SocketException e) {
